Restructures form elements to match AA's styles
ActiveAdmin uses Formtastic to build and style their forms, which uses a
specific set of markup elements and classes to provide a consistent look
and feel. This updates the elements of a few custom forms (inventory
filter, donation add items) so their labels and control sizes match all
of the other forms.