Fixes inventory count totals on dashboard table
Queries Items in inventory, grouping by item, inventory, and summing up
quantity (to account for possible duplicates). Then the data is
restructed to be rendered in the inventory dashboard summary partial.
This also optimizes the number of loops it takes to build the data.

However there is a small cost of looping over the inventories array
multiple times during rendering to pull the data out of items hash.
Maybe there is room for improvement there.