Checking in migration to create a new transfer model and the updated schema.
The transfer model is a fairly close match to the ticket model but it has a belongs_to :from and belongs_to :to mapping on to the inventory model.

This isn't the best solution but since we promised the stakeholder a date of Jul 31 for MVP I'm happy just getting something shipped right now. :)