Updated inventory with a method move_inventory! to allow the transfering from one location to another.
The move_inventory was plagarized... er... emulated from the distribute! method with a few items added to reference the receiving location.

The transfer model is a close approximation of the ticket model and the the transfer active admin page is also a close approximation of the ticket page.

I'll have to dry these up.